backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
sqlalchemy>=2.0.25

pydantic>=2.12.3
pydantic-settings>=2.11.0

httpx>=0.28.1
python-dotenv>=1.1.1

apscheduler>=3.10.4
